requirements.txt
flask==2.3.3
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
diffusers>=0.24.0
accelerate>=0.24.0
safetensors>=0.4.0
pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
opencv-python>=4.8.0
xformers>=0.0.22
